+++ src/Pim/Attribute/AttributeService.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Wizaplace\SDK\Pim\Product\Attribute;
+namespace Wizaplace\SDK\Pim\Attribute;
 
 use Wizaplace\SDK\AbstractService;
 
@@ -36,5 +36,20 @@ final class AttributeService extends AbstractService
         $data = $this->client->get("products/${productId}/features/${attributeId}");
 
         return ProductAttribute::build($data);
+    }
+
+    /**
+     * @return Attribute[]
+     * @throws \Wizaplace\SDK\Authentication\AuthenticationRequired
+     */
+    public function getCategoryAttributes(int $categoryId): array
+    {
+        $this->client->mustBeAuthenticated();
+
+        $data = $this->client->get("categories/${categoryId}/features");
+
+        return array_map(static function (array $attributeData): Attribute {
+            return new Attribute($attributeData);
+        }, $data);
     }
 }