+++ tests/Pim/Attribute/AttributeServiceTest.php
@@ -5,14 +5,15 @@
  */
 declare(strict_types=1);
 
-namespace Wizaplace\SDK\Tests\Pim\Product\Attribute;
+namespace Wizaplace\SDK\Tests\Pim\Attribute;
 
-use Wizaplace\SDK\Pim\Product\Attribute\AttributeService;
-use Wizaplace\SDK\Pim\Product\Attribute\AttributeType;
-use Wizaplace\SDK\Pim\Product\Attribute\AttributeVariant;
-use Wizaplace\SDK\Pim\Product\Attribute\ProductAttribute;
-use Wizaplace\SDK\Pim\Product\Attribute\ProductAttributeValue;
-use Wizaplace\SDK\Pim\Product\Attribute\ProductAttributeVariants;
+use Wizaplace\SDK\Pim\Attribute\Attribute;
+use Wizaplace\SDK\Pim\Attribute\AttributeService;
+use Wizaplace\SDK\Pim\Attribute\AttributeType;
+use Wizaplace\SDK\Pim\Attribute\AttributeVariant;
+use Wizaplace\SDK\Pim\Attribute\ProductAttribute;
+use Wizaplace\SDK\Pim\Attribute\ProductAttributeValue;
+use Wizaplace\SDK\Pim\Attribute\ProductAttributeVariants;
 use Wizaplace\SDK\Tests\ApiTestCase;
 
 class AttributeServiceTest extends ApiTestCase
@@ -36,6 +37,29 @@ class AttributeServiceTest extends ApiTestCase
                 $variants = $attribute->getSelectedVariants();
                 $this->assertContainsOnly(AttributeVariant::class, $variants);
 
+                foreach ($variants as $variant) {
+                    $this->assertNotEmpty($variant->getName());
+                    $this->assertGreaterThan(0, $variant->getId());
+                }
+            }
+        }
+    }
+
+    public function testGetCategoryAttributes()
+    {
+        $attributes = $this->buildAttributeService()->getCategoryAttributes(6);
+
+        $this->assertContainsOnly(Attribute::class, $attributes);
+        $this->assertGreaterThanOrEqual(7, count($attributes));
+
+        foreach ($attributes as $attribute) {
+            $this->assertNotEmpty($attribute->getName());
+            $this->assertGreaterThan(0, $attribute->getId());
+            $this->assertContainsOnly('int', $attribute->getCategoriesPath());
+            $this->assertInstanceOf(AttributeType::class, $attribute->getType());
+            $variants = $attribute->getVariants();
+            if ($variants !== null) {
+                $this->assertContainsOnly(AttributeVariant::class, $variants);
                 foreach ($variants as $variant) {
                     $this->assertNotEmpty($variant->getName());
                     $this->assertGreaterThan(0, $variant->getId());